rename conda/core/repodata.py to conda/core/subdir_data.py
