+++ conda/cli/main_info.py
@@ -22,7 +22,7 @@ from ..base.context import conda_in_private_env, context, sys_rc_path, user_rc_p
 from ..common.compat import iteritems, itervalues, on_win, text_type
 from ..common.url import mask_anaconda_token
 from ..core.envs_manager import env_name
-from ..core.repodata import SubdirData
+from ..core.subdir_data import SubdirData
 from ..models.channel import all_channel_urls, offline_keep
 from ..models.match_spec import MatchSpec
 from ..utils import human_bytes
