+++ conda/core/subdir_data.py
